Fix element height and watch active items.
The items in the child scope need updating from the collection (which must be evaluated each time).

Some attempt is made to fix the height of each element - this will fail in the face of padding or borders!

Also found a bug in scrolling backwards - wasn't deleting rows.

Added a new demo to exercise big lists.