More special casing for tables.
Tables can't be the content as we need to size them explicitly without
putting all the rows in. It's therefore necessary to have an extra div
around a table. It's also necessary to set the row height explicitly if
IE is to be able to pick it up.

Not proud of this one!